backspace clears the log
I'm not good enough at this to get it to backspace the log and backspace selections or understand selections via mouse. So when backspace is used, it delete's one character. 

It also is ignoring [SHIFT] and [BACKSPACE] and thus not entering those into the log. 
What this accomplishes is a slightly cleaner log, not by much though. I had much less success in this than I thought I would but for my purposes this works well enough.